feat: IDE-12-ContainerService 단위 테스트를 위한 ContainerUtil 인터페이스 분리
